MNIST dataset code fully tested

+++ torch-extension/include/datasets/mnist.h
@@ -40,7 +40,7 @@ namespace torch::ext::data::datasets {
 
     private :
         std::vector<torch::Tensor> data; // Store image data as tensors
-        std::vector<int64_t> labels;      // Store labels
+        std::vector<uint8_t> labels;      // Store labels
         std::string url = "https://ossci-datasets.s3.amazonaws.com/mnist/";
         fs::path root;
         fs::path dataset_path;
